Add files via upload
Given a 2D design space of 150cm long and 50cm tall, a rigid bar of a fixed length (yourchoice), 3 pin supports of which two need to be mounted on the ground and a linear actuator (pick from this online catalog, use max force values only), design a frame/mechanism to lift the maximum possible weight to the highest possible height. Assume all the supports and bar/actuator are rigid.